requirements.txt
streamlit
pandas
plotly
mysql-connector-python
cassandra-driver
scikit-learn

